refactor: keyboard logic in other file
